requirements.txt
# ETL Pipeline Dependencies
paramiko==3.3.1
google-cloud-storage==2.10.0
google-cloud-bigquery==3.13.0
google-cloud-dataproc==5.8.0
google-cloud-composer==1.9.0
functions-framework==3.5.0
apache-airflow==2.7.1
apache-airflow-providers-google==10.8.0
pyspark==3.5.0
polars==0.19.12
pandas==2.0.3
sqlalchemy==2.0.21
pytest==7.4.2
pytest-cov==4.1.0
pyyaml==6.0.1
python-dotenv==1.0.0
requests==2.31.0
